+++ CHANGELOG.md
@@ -3,11 +3,18 @@
 The Messenger library follows [SemVer](http://semver.org/).
 
 ## 3.x
-**Changelog** (since [`3.3.0`](https://github.com/ker0x/messenger/compare/3.2.0...3.3.0))
+
+**Changelog** (since [`3.2.0`](https://github.com/ker0x/messenger/compare/3.2.0...3.3.0))
+
 - 3.3.0 (2019-11)
-    - Support of `image_aspect_ratio` for Generic Template
+    - Update API version to `v5.0`
+    - Add Support of `image_aspect_ratio` property for Generic Template (Thanks to @BFoucher)
+    - Add support of `ice_breakers` and `home_url` properties for Profile API
+    - Add support of `reply_to` property for Message callback
+    - Add support of `Reaction` event
+    - Properties `payment_settings` and `target_audience` are now deprecated
     
-**Changelog** (since [`3.2.0`](https://github.com/ker0x/messenger/compare/3.1.1...3.2.0))
+**Changelog** (since [`3.1.1`](https://github.com/ker0x/messenger/compare/3.1.1...3.2.0))
 
 - 3.2.0 (2019-08)
     - Update API version to `v3.3`